+++ packages/orq-rc/src/orq_ai_sdk/models/createpromptsnippetop.py
@@ -518,8 +518,6 @@ class CreatePromptSnippetMetadata(BaseModel):
 
 class CreatePromptSnippetRequestBodyTypedDict(TypedDict):
     key: str
-    display_name: str
-    r"""The prompt snippet’s name, meant to be displayable in the UI."""
     prompt_config: PromptConfigTypedDict
     r"""A list of messages compatible with the openAI schema"""
     path: str
@@ -531,9 +529,6 @@ class CreatePromptSnippetRequestBodyTypedDict(TypedDict):
 
 class CreatePromptSnippetRequestBody(BaseModel):
     key: str
-
-    display_name: str
-    r"""The prompt snippet’s name, meant to be displayable in the UI."""
 
     prompt_config: PromptConfig
     r"""A list of messages compatible with the openAI schema"""
@@ -1724,8 +1719,6 @@ class VersionsTypedDict(TypedDict):
     r"""Prompt version model returned from the API"""
 
     id: str
-    display_name: str
-    r"""The prompt snippet’s name, meant to be displayable in the UI."""
     prompt_config: CreatePromptSnippetPromptSnippetsPromptConfigTypedDict
     r"""A list of messages compatible with the openAI schema"""
     metadata: CreatePromptSnippetPromptSnippetsResponseMetadataTypedDict
@@ -1740,9 +1733,6 @@ class Versions(BaseModel):
     r"""Prompt version model returned from the API"""
 
     id: Annotated[str, pydantic.Field(alias="_id")]
-
-    display_name: str
-    r"""The prompt snippet’s name, meant to be displayable in the UI."""
 
     prompt_config: CreatePromptSnippetPromptSnippetsPromptConfig
     r"""A list of messages compatible with the openAI schema"""
@@ -1796,8 +1786,6 @@ class CreatePromptSnippetResponseBodyTypedDict(TypedDict):
     owner: OwnerTypedDict
     domain_id: str
     key: str
-    display_name: str
-    r"""The prompt snippet’s name, meant to be displayable in the UI."""
     prompt_config: CreatePromptSnippetPromptConfigTypedDict
     r"""A list of messages compatible with the openAI schema"""
     metadata: CreatePromptSnippetPromptSnippetsMetadataTypedDict
@@ -1824,9 +1812,6 @@ class CreatePromptSnippetResponseBody(BaseModel):
 
     key: str
 
-    display_name: str
-    r"""The prompt snippet’s name, meant to be displayable in the UI."""
-
     prompt_config: CreatePromptSnippetPromptConfig
     r"""A list of messages compatible with the openAI schema"""
 
@@ -1846,7 +1831,7 @@ class CreatePromptSnippetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-05T12:50:57.727Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-05T18:57:06.512Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")