+++ packages/orq-rc/src/orq_ai_sdk/models/findonepromptsnippetop.py
@@ -1177,8 +1177,6 @@ class FindOnePromptSnippetVersionsTypedDict(TypedDict):
     r"""Prompt version model returned from the API"""
 
     id: str
-    display_name: str
-    r"""The prompt snippet’s name, meant to be displayable in the UI."""
     prompt_config: FindOnePromptSnippetPromptSnippetsPromptConfigTypedDict
     r"""A list of messages compatible with the openAI schema"""
     metadata: FindOnePromptSnippetPromptSnippetsMetadataTypedDict
@@ -1193,9 +1191,6 @@ class FindOnePromptSnippetVersions(BaseModel):
     r"""Prompt version model returned from the API"""
 
     id: Annotated[str, pydantic.Field(alias="_id")]
-
-    display_name: str
-    r"""The prompt snippet’s name, meant to be displayable in the UI."""
 
     prompt_config: FindOnePromptSnippetPromptSnippetsPromptConfig
     r"""A list of messages compatible with the openAI schema"""
@@ -1249,8 +1244,6 @@ class FindOnePromptSnippetResponseBodyTypedDict(TypedDict):
     owner: FindOnePromptSnippetOwnerTypedDict
     domain_id: str
     key: str
-    display_name: str
-    r"""The prompt snippet’s name, meant to be displayable in the UI."""
     prompt_config: FindOnePromptSnippetPromptConfigTypedDict
     r"""A list of messages compatible with the openAI schema"""
     metadata: FindOnePromptSnippetMetadataTypedDict
@@ -1277,9 +1270,6 @@ class FindOnePromptSnippetResponseBody(BaseModel):
 
     key: str
 
-    display_name: str
-    r"""The prompt snippet’s name, meant to be displayable in the UI."""
-
     prompt_config: FindOnePromptSnippetPromptConfig
     r"""A list of messages compatible with the openAI schema"""
 
@@ -1299,7 +1289,7 @@ class FindOnePromptSnippetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-05T12:50:57.727Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-05T18:57:06.512Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")